Nuove modifiche 01/09/2016

+++ src/main/java/com/websystique/springmvc/dao/AbstractDao.java
@@ -33,7 +33,7 @@ public abstract class AbstractDao<PK extends Serializable, T> {
 	}
 
 	@SuppressWarnings("unchecked")
-	public T getByKey(BigInteger id) {
+	public T getByKey(Integer id) {
 		return (T) getSession().get(persistentClass, id);
 	}
